+++ 02_daemonset.tf
@@ -217,10 +217,10 @@ resource "kubernetes_daemon_set_v1" "daemonset" {
               run_as_user               = init_container.value.securityContext.runAsUser
 
               dynamic "seccomp_profile" {
-                for_each = init_container.value.securityContext.seccomp_profile.type != null ? [init_container.value.securityContext.seccomp_profile] : []
+                for_each = init_container.value.securityContext.seccompProfile.type != null ? [init_container.value.securityContext.seccompProfile] : []
                 content {
-                  localhost_profile = init_container.value.securityContext.seccomp_profile.localhost_profile
-                  type              = init_container.value.securityContext.seccomp_profile.type
+                  localhost_profile = init_container.value.securityContext.seccompProfile.localhostProfile
+                  type              = init_container.value.securityContext.seccompProfile.type
                 }
               }
             }
@@ -470,10 +470,10 @@ resource "kubernetes_daemon_set_v1" "daemonset" {
               run_as_user               = container.value.securityContext.runAsUser
 
               dynamic "seccomp_profile" {
-                for_each = container.value.securityContext.seccomp_profile.type != null ? [container.value.securityContext.seccomp_profile] : []
+                for_each = container.value.securityContext.seccompProfile.type != null ? [container.value.securityContext.seccompProfile] : []
                 content {
-                  localhost_profile = container.value.securityContext.seccomp_profile.localhost_profile
-                  type              = container.value.securityContext.seccomp_profile.type
+                  localhost_profile = container.value.securityContext.seccompProfile.localhostProfile
+                  type              = container.value.securityContext.seccompProfile.type
                 }
               }
             }